requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
sqlmodel==0.0.21
aiosqlite==0.20.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.1
streamlit==1.37.1
openai==1.37.0
requests==2.32.3